Tests/LibWeb/Text/expected/canvas/webgl-capability-only-extension-object.txt

OES_texture_float_linear supported: true
OES_texture_float_linear object: object
OES_texture_float_linear same object: true
OES_texture_float_linear own properties: 0
